Set the ObjectTracker::tracked_object_ptr_ to nullptr when the tracked object is destroyed.
Not really necessary, but if a weak_ptr<Tracker> is converted to
a shared_ptr and while that shared_ptr is alive the the tracked
object is destroyed, then accessing another weak_ptr (or that
shared_ptr) will access the memory of the deleted object.

This way that at least won't go by unnoticed.