Bug fix.
The author of enchantum wrote me (regarding the comment):

Well it doesn't do what is advertised it always calls enchantum::to_string
even if there is adl for it because enchantum::to_string is a functor not
a function it doesn't participate in adl. The fix is to wrap enchantum::to_string.